dataloader/app/requirements.txt
pandas
uuid
cassandra-driver